debugging FSAssignmentController POST method
change conditional logic to check for null value
also removed a comment from assign jsp
both changes made no difference

print statements show an unpopulated animal. animalID and
feedingScheduleID=0, all other fields null

+++ src/com/examples/ezoo/controllers/FSAssignmentController.java
@@ -111,7 +111,7 @@ public class FSAssignmentController {
 			}
 			
 			// unassignment logic
-			if (animal.getFeedingScheduleID() > 0) {
+			if (animal.getFeedingScheduleID() != null && animal.getFeedingScheduleID() > 0) { // not null and not zero
 				FSDAO.removeFeedingSchedule(animal);
 //				request.getSession().setAttribute("message",  "Feeding schedule successfully removed");
 				model.addAttribute("message",  "Feeding schedule successfully removed");
